+++ header.php
@@ -19,3 +19,4 @@
 	<body <?php body_class(); ?>>
 		<?php do_action( 'wp_body_open' ); ?>
 		<?php get_template_part( 'template-parts/header', 'main' ); ?>
+		<?php get_template_part( 'template-parts/menu-overlay' ); ?>